LayoutTests/platform/ios/media/video-source-before-src-expected.txt

Append source element before setting src attribute.
Platforms that support setRequiresUserGestureForVideoPlayback will have the expected, empty, currentSrc. Platforms that don't should have the unexpected "src.mp4".

EXPECTED (relativeURL(video.currentSrc) == '') OK
